2023_06_28_0110-wmtag_memo_完善了10号脚本

+++ 10.wmscript_init_this.sh
@@ -14,11 +14,11 @@ case `uname -s` in
 "Linux")
 	PLATFORM="linux"
 	# wmtag_memo_我修改了这里_开始
-	apt update 
-	apt install -y lsof net-tools direnv strace 
-	apt install -y libbsd-dev
-	sudo cp include/apue.h /usr/include/
-	sudo cp lib/error.c /usr/include/ 
+	[[ -f $(which cloudstudio) ]] && apt update 
+	[[ -f $(which cloudstudio) ]] && apt install -y lsof net-tools direnv strace 
+	[[ -f $(which cloudstudio) ]] && apt install -y libbsd-dev
+	[[ -f $(which cloudstudio) ]] && cp include/apue.h /usr/include/
+	[[ -f $(which cloudstudio) ]] && cp lib/error.c /usr/include/ 
 	# wmtag_memo_我修改了这里_结束
 	;;
 "Darwin")
